src/backend/pyproject.toml
[project]
name = "open-trainaa"
version = "1.0.4"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite==0.21.0",
    "annotated-types==0.7.0",
    "anthropic==0.64.0",
    "anyio==4.10.0",
    "appnope==0.1.4",
    "apscheduler>=3.11.0",
    "asttokens==3.0.0",
    "asyncpg>=0.30.0",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "comm==0.2.3",
    "contourpy==1.3.3",
    "cryptography>=45.0.7",
    "cycler==0.12.1",
    "dateparser>=1.2.2",
    "debugpy==1.8.16",
    "decorator==5.2.1",
    "deprecated==1.2.18",
    "deprecation==2.1.0",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "dydantic==0.0.8",
    "email-validator==2.2.0",
    "executing==2.2.0",
    "fastapi==0.116.1",
    "fastapi-cli==0.0.8",
    "fastapi-cloud-cli==0.1.5",
    "fastmcp==2.12.2",
    "fit-tool",
    "fitparse==1.2.0",
    "fonttools==4.59.1",
    "garmin-fit-sdk==21.178.0",
    "garth==0.5.17",
    "h11==0.16.0",
    "h2==4.2.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "hyperframe==6.1.0",
    "idna==3.10",
    "iniconfig==2.1.0",
    "ipykernel==6.30.1",
    "ipython==9.4.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "kiwisolver==1.4.9",
    "langchain==0.3.27",
    "langchain-anthropic==0.3.19",
    "langchain-core==0.3.74",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai==0.3.30",
    "langchain-text-splitters==0.3.9",
    "langfuse>=3.3.4",
    "langgraph==0.6.6",
    "langgraph-checkpoint==2.1.1",
    "langgraph-checkpoint-postgres==2.0.23",
    "langgraph-checkpoint-sqlite==2.0.11",
    "langgraph-prebuilt==0.6.4",
    "langgraph-sdk==0.2.6",
    "langmem==0.0.29",
    "langsmith==0.4.15",
    "limits==5.5.0",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.2",
    "matplotlib==3.10.5",
    "matplotlib-inline==0.1.7",
    "mcp>=1.14.0",
    "mdurl==0.1.2",
    "nest-asyncio==1.6.0",
    "numpy==2.3.2",
    "oauthlib==3.3.1",
    "openai==1.100.2",
    "orjson==3.11.2",
    "ormsgpack==1.10.0",
    "packaging==25.0",
    "pandas==2.3.2",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pillow==11.3.0",
    "platformdirs==4.3.8",
    "pluggy==1.6.0",
    "postgrest==1.1.1",
    "prompt-toolkit==3.0.51",
    "psutil==7.0.0",
    "psycopg-pool==3.2.6",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary]>=3.2.9",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pygments==2.19.2",
    "pyjwt[crypto]==2.10.1",
    "pyparsing==3.2.3",
    "pytest==8.4.1",
    "pytest-asyncio==1.1.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "pyzmq==27.0.2",
    "realtime==2.7.0",
    "redis[hiredis]>=7.0.1",
    "regex==2025.7.34",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.1.0",
    "rich-toolkit==0.15.0",
    "rignore==0.6.4",
    "ruff>=0.13.1",
    "scipy==1.16.1",
    "seaborn==0.13.2",
    "sentry-sdk[fastapi]==2.35.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "slowapi==0.1.9",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.43",
    "sqlite-vec==0.1.6",
    "stack-data==0.6.3",
    "starlette==0.47.2",
    "storage3==0.12.1",
    "strenum==0.4.15",
    "stripe>=13.2.0",
    "supabase==2.18.1",
    "supabase-auth==2.12.3",
    "supabase-functions==0.10.1",
    "tenacity==9.1.2",
    "tiktoken==0.11.0",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "trustcall==0.0.39",
    "typer==0.16.1",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "wcwidth==0.2.13",
    "websockets==15.0.1",
    "wrapt==1.17.3",
    "xxhash==3.5.0",
    "zstandard==0.24.0",
]

[tool.uv.sources]
fit-tool = { path = "python_fit_tool_jnkue", editable = false }

[tool.uv]
dev-dependencies = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]
